// src/Ambassador/Models/ProgramDetails.ts
export default class ProgramDetails {
  id: number = -1;
  email: string = "";
  max_reward: string = "";
  title: string = "";
  description: string = "";
  ambassador_program_img: string = "";
  ambassador_img: string = "";
  crypto_wallet: string = "";
  paypal_email: string = "";
  instagram_link: string = "";
  twitter_link: string = "";
  website_link: string = "";
  community_channel: string = "";
  community_group: string = "";
  linkedin_link: string = "";
  total_points: string = "";
  lifetime_points: string = "";
  mailing_address: string = "";
  mailing_city: string = "";
  mailing_state: string = "";
  mailing_country: string = "";
  status: boolean = false;
  createdAt: string = "";
  updatedAt: string = "";
  ambassador_id: number = -1;
  program_id: number = -1;
  program_title: string = "";
  program_email: string = "";
  program_img: string = "";
  expiry_date: string = "";
  timeline: number = -1;
  leaderboard: number = -1;
  amb_reward: boolean = true;
}