// src/Ambassador/Models/CampaignDetails.ts
export default class CampaignDetails {
  id: number = -1;
  title: string = "";
  description: string = "";
  start_date: string = "";
  end_date: string = "";
  campaign_img: string = "";
  points: number = -1;
  status: string = "";
  createdAt: string = "";
  updatedAt: string = "";
  program_id: number = -1;
  automatic: number = -1;
  platform: number = -1;
  completion_status: number = 0;
  amb_prog_id: number = -1;
  custom_camp_id: number = -1;
  difference: number = -1;
}